fix or silence compiler warnings about unused parameters, uninitialized local variables, and narrowing
(cherry picked from commit 9e37dba66ea672d11567e75b99f0ac38c56c9891)

+++ include/boost/numeric/odeint/integrate/detail/integrate_const.hpp
@@ -80,7 +80,7 @@ size_t integrate_const(
 
     Time time = start_time;
     const Time time_step = dt;
-    int real_steps = 0;
+    size_t real_steps = 0;
     int step = 0;
     
     while( less_eq_with_sign( static_cast<Time>(time+time_step) , end_time , dt ) )
